Rename TangledCommitDetection class into TangledChangeDetector since it can also analyse diffs belonging to more than one single commit
