+++ src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/TangledChangeDetector.java
@@ -7,7 +7,7 @@ import de.uhd.ifi.se.decision.management.jira.model.git.Diff;
  * Interface for the estimation whether a {@link Diff} of {@ChangedFile}s
  * contains wrong links, i.e., is tangled.
  */
-public interface TangledCommitDetection {
+public interface TangledChangeDetector {
 
 	/**
 	 * Currently, this function calls {@link #calculatePackageDistances(Diff)}.
