+++ src/main/java/de/uhd/ifi/se/decision/management/jira/extraction/impl/TangledChangeDetectorImpl.java
@@ -2,11 +2,15 @@ package de.uhd.ifi.se.decision.management.jira.extraction.impl;
 
 import java.util.List;
 
-import de.uhd.ifi.se.decision.management.jira.extraction.TangledCommitDetection;
+import de.uhd.ifi.se.decision.management.jira.extraction.TangledChangeDetector;
 import de.uhd.ifi.se.decision.management.jira.model.git.ChangedFile;
 import de.uhd.ifi.se.decision.management.jira.model.git.Diff;
 
-public class TangledCommitDetectionImpl implements TangledCommitDetection {
+/**
+ * Class for the estimation whether a {@link Diff} of {@ChangedFile}s
+ * contains wrong links, i.e., is tangled.
+ */
+public class TangledChangeDetectorImpl implements TangledChangeDetector {
 
 	@Override
 	public void estimateWhetherChangedFilesAreCorrectlyIncludedInDiff(Diff diff) {
