+++ src/test/java/de/uhd/ifi/se/decision/management/jira/extraction/tangledchangedetector/TestTangledChangeDetector.java
@@ -1,4 +1,4 @@
-package de.uhd.ifi.se.decision.management.jira.extraction.tangledcommitdetection;
+package de.uhd.ifi.se.decision.management.jira.extraction.tangledchangedetector;
 
 import static org.junit.Assert.assertEquals;
 
@@ -9,24 +9,24 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import de.uhd.ifi.se.decision.management.jira.extraction.TangledCommitDetection;
+import de.uhd.ifi.se.decision.management.jira.extraction.TangledChangeDetector;
 import de.uhd.ifi.se.decision.management.jira.extraction.gitclient.TestSetUpGit;
-import de.uhd.ifi.se.decision.management.jira.extraction.impl.TangledCommitDetectionImpl;
+import de.uhd.ifi.se.decision.management.jira.extraction.impl.TangledChangeDetectorImpl;
 import de.uhd.ifi.se.decision.management.jira.model.git.ChangedFile;
 import de.uhd.ifi.se.decision.management.jira.model.git.Diff;
 import de.uhd.ifi.se.decision.management.jira.model.git.impl.DiffImpl;
 
-public class TestTangledCommitDetection extends TestSetUpGit {
+public class TestTangledChangeDetector extends TestSetUpGit {
 
 	private Diff diffForJiraIssue;
 	private Diff diffForCommit;
 
-	private TangledCommitDetection tangledCommitDetection;
+	private TangledChangeDetector tangledCommitDetection;
 
 	@Before
 	public void setUp() {
 		super.setUp();
-		tangledCommitDetection = new TangledCommitDetectionImpl();
+		tangledCommitDetection = new TangledChangeDetectorImpl();
 		diffForCommit = TestDiff.createDiff(mockJiraIssueForGitTestsTangledSingleCommit);
 		diffForJiraIssue = TestDiff.createDiff(mockJiraIssueForGitTestsTangled);
 	}
